requirements.txt
gTTS==1.2.2
Flask-API==1.0
Flask-Cors==3.0.3
Flask-APScheduler==1.8.0
pychromecast==2.0.0
unicode_slugify==0.1.3
webargs==2.0.0
arrow==0.12.1
requests==2.18.4
youtube-dl==2018.3.14
